# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/example/getAuthorizationHeader": {
            "get": {
                "description": "Get 请求, 请求参数是 URL 的一部分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Get 请求, 请求参数是 URL 的一部分",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/getExtendAttribute": {
            "get": {
                "description": "Get 请求, 请求参数在 URL 中携带, 对参数配置了校验规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Get 请求, 请求参数在 URL 中携带, 对参数配置了校验规则",
                "parameters": [
                    {
                        "enum": [
                            "A",
                            "B",
                            "C"
                        ],
                        "type": "string",
                        "description": "字符串类型,参数必须是下面下拉列表中的值",
                        "name": "enumString",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "整数类型, 参数必须是下面下拉列表中的值",
                        "name": "enumInt",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1.1,
                            1.2,
                            1.3
                        ],
                        "type": "number",
                        "description": "浮点数类型, 参数必须是下面下拉列表中的值",
                        "name": "enumNumber",
                        "in": "query"
                    },
                    {
                        "maxLength": 10,
                        "minLength": 5,
                        "type": "string",
                        "description": "验证规则:字符串最小长度 5, 最大长度 10",
                        "name": "string",
                        "in": "query"
                    },
                    {
                        "maximum": 10,
                        "minimum": 1,
                        "type": "integer",
                        "description": "验证部规则: 整数最小值为 1, 最大值为 10",
                        "name": "int",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "A",
                        "description": "提供默认的字符串在编辑框中",
                        "name": "default",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/getWithPathVar/groups/{group_id}/accounts/{account_id}": {
            "get": {
                "description": "Get 请求, 请求参数是 URL 的一部分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Get 请求, 请求参数是 URL 的一部分",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组 ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账号 ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/getWithUrlQueryString": {
            "get": {
                "description": "使用 Get 请求, 在 URL 上携带参数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "使用 Get 请求, 在 URL 上携带参数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postAuthorizationHeader": {
            "post": {
                "description": "Post 请求, 参数在 Header 中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 参数在 Header 中",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postExtendAttribute": {
            "post": {
                "description": "Post 请求, 请求参数在 URL 中携带, 对参数配置了校验规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 请求参数在 URL 中携带, 对参数配置了校验规则",
                "parameters": [
                    {
                        "enum": [
                            "A",
                            "B",
                            "C"
                        ],
                        "type": "string",
                        "description": "字符串类型,参数必须是下面下拉列表中的值",
                        "name": "enumString",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "整数类型, 参数必须是下面下拉列表中的值",
                        "name": "enumInt",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1.1,
                            1.2,
                            1.3
                        ],
                        "type": "number",
                        "description": "浮点数类型, 参数必须是下面下拉列表中的值",
                        "name": "enumNumber",
                        "in": "query"
                    },
                    {
                        "maxLength": 10,
                        "minLength": 5,
                        "type": "string",
                        "description": "验证规则:字符串最小长度 5, 最大长度 10",
                        "name": "string",
                        "in": "query"
                    },
                    {
                        "maximum": 10,
                        "minimum": 1,
                        "type": "integer",
                        "description": "验证部规则: 整数最小值为 1, 最大值为 10",
                        "name": "int",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "A",
                        "description": "提供默认的字符串在编辑框中",
                        "name": "default",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postFormData": {
            "post": {
                "description": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "parameters": [
                    {
                        "type": "string",
                        "description": "前端发送的表单中的 name 参数",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "前端发送的表单中的 age 参数",
                        "name": "age",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postJsonData": {
            "post": {
                "description": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 发送 Json 数据, 参数在消息体中",
                "parameters": [
                    {
                        "description": "前端发送的 Json 对象",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postUrlEncodedFormData": {
            "post": {
                "description": "Post 请求, 发送表单数据, 表单类型为: x-www-form-urlencoded, 参数在消息体中",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 发送表单数据, 表单类型为: x-www-form-urlencoded, 参数在消息体中",
                "parameters": [
                    {
                        "type": "string",
                        "description": "前端发送的表单中的 name 参数",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "前端发送的表单中的 age 参数",
                        "name": "age",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/example/postWithPathVariable/groups/{group_id}/accounts/{account_id}": {
            "post": {
                "description": "Post 请求, 请求参数是 URL 的一部分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exampleApi"
                ],
                "summary": "Post 请求, 请求参数是 URL 的一部分",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组 ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账号 ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "is_delete": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "tel": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "swagger.Http400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "swagger.Http404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Page Not Found"
                }
            }
        },
        "swagger.Http500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "服务器内部错误"
                }
            }
        },
        "swagger.HttpOk": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "接口调用成功"
                }
            }
        }
    }
}